Rename DatabaseExecutor on RichDBIO
